[Fixed] OC-487 #comment fixed function that deletes folders during off update process #time 20m

+++ src/backend/src/main/java/com/becon/opencelium/backend/application/assistant/AssistantServiceImp.java
@@ -359,7 +359,6 @@ public class AssistantServiceImp implements ApplicationService {
 //            unzipFolder(oc, target);
     }
 
-
     private void deleteDir(File[] files, Stack<String> pathParts) {
         try {
             for (File file : files) {
@@ -368,7 +367,10 @@ public class AssistantServiceImp implements ApplicationService {
                     deleteDir(file.listFiles(), pathParts);
                     continue;
                 }
-                FileUtils.deleteDirectory(file);
+                if (file.isDirectory()) {
+                    FileUtils.deleteDirectory(file);
+                }
+                Files.delete(file.toPath());
             }
         } catch (Exception e) {
             throw new RuntimeException(e);
